pyproject.toml
[project]
name = "fest-ticketing-backend"
version = "0.1.0"
description = "A simple E-Ticketing System for concerts and events"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]<1.0.0,>=0.115.0",
    "python-multipart<1.0.0,>=0.0.12",
    "email-validator<3.0.0.0,>=2.2.0",
    "passlib[bcrypt]<2.0.0,>=1.7.4",
    "tenacity<9.0.0,>=8.2.3",
    "pydantic<3.0.0,>=2.9.2",
    "emails<1.0,>=0.6",
    "httpx<1.0.0,>=0.27.2",
    # Pin bcrypt until passlib supports the latest
    "bcrypt==4.2.0",
    "pydantic-settings>=2.5.2,<3.0.0",
    "pyjwt>=2.9.0,<3.0.0",
    "opencv-python>=4.10.0.84",
    "ruff>=0.6.8",
    "google-auth>=2.35.0",
    "python-jose>=3.3.0",
    "redis[hiredis]>=5.2.0",
    "jinja2>=3.1.4",
    "alembic>=1.14.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "cloudinary>=1.41.0",
    "psycopg[binary,pool]>=3.2.3",
    "pyotp>=2.9.0",
    "keras-facenet>=0.3.2",
    "mediapipe>=0.10.18",
    "tensorflow>=2.18.0",
    "scikit-learn>=1.5.2",
    "apscheduler>=3.11.0",
]

[tool.uv]
dev-dependencies = [
    "pytest<9.0.0,>=8.3.3",
    "mypy<2.0.0,>=1.11.2",
    "ruff<1.0.0,>=0.6.8",
    "types-passlib<2.0.0.0,>=1.7.7.20240819",
    "coverage<8.0.0,>=7.6.1",
    "faker>=30.8.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = ["/.github", "/docs", "/app/tests"]

[tool.hatch.build.targets.wheel]
packages = ["app/*"]

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py310"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "W191", # indentation contains tabs
    "B904", # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
